- Created a package for invoice header stuff; moved there DatiTrasmissioneType, FatturaElettronicaHeaderType and IdFiscaleType - Started documenting those classes
