+++ src/main/java/com/github/siboxd/fatturapa/model/DatiAnagraficiCessionarioType.java
@@ -1,5 +1,7 @@
 package com.github.siboxd.fatturapa.model;
 
+import com.github.siboxd.fatturapa.model.invoiceheader.IdFiscaleType;
+
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
